feat: AI 코멘트 모달을 일기 작성페이지로 이동 feature/TODAK-213/update-testcase
